requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alchemy-mock==0.4.3
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cloud-sql-python-connector==0.7.0
cryptography==37.0.4
frozenlist==1.3.0
google-auth==2.9.1
greenlet==1.1.2
idna==3.3
multidict==6.0.2
pg8000==1.29.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyOpenSSL==22.0.0
python-decouple==3.6
requests==2.28.1
rsa==4.8
scramp==1.4.1
six==1.16.0
SQLAlchemy==1.4.39
toml==0.10.2
tornado==6.2
urllib3==1.26.10
yarl==1.7.2
